Remove harmonizer & supergraph (#504)
We're now building harmonizer & supergraph internally from a different
repo. This PR removes them, as well as their `xtask` residue from this
repo.

+++ apollo-federation-types/src/build_plugin/build_message.rs
@@ -10,7 +10,7 @@ pub enum BuildMessageLevel {
     Error,
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+#[derive(Debug, Default, Clone, Serialize, Deserialize, PartialEq, Eq)]
 /// BuildLocation represents the location of a build message in the GraphQLDoucment
 /// New fields added to this struct must be optional in order to maintain
 /// backwards compatibility with old versions of Rover
@@ -28,17 +28,17 @@ pub struct BuildMessageLocation {
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 pub struct BuildMessagePoint {
-    pub start: Option<u32>,
-    pub end: Option<u32>,
-    pub column: Option<u32>,
-    pub line: Option<u32>,
+    pub start: Option<usize>,
+    pub end: Option<usize>,
+    pub column: Option<usize>,
+    pub line: Option<usize>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
 #[serde(rename_all = "camelCase")]
 /// BuildMessages contains the log output of a build
 /// New fields added to this struct must be optional in order to maintain
-/// backwards compatibility with old versions of Rover
+/// backwards compatibility
 pub struct BuildMessage {
     pub level: BuildMessageLevel,
     pub message: String,
